move dependency solver tests to their own file
