+++ nuget/helpers/lib/NuGetUpdater/NuGetUpdater.Core.Test/Utilities/MSBuildHelperTests.cs
@@ -678,1049 +678,6 @@ public class MSBuildHelperTests : TestBase
         ];
     }
 
-    #region
-    // Updating root package
-    // CS-Script Code to 2.0.0 requires its dependency Microsoft.CodeAnalysis.CSharp.Scripting to be 3.6.0 and its transitive dependency Microsoft.CodeAnalysis.Common to be 3.6.0
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewUpdatingTopLevelPackage(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="CS-Script.Core" Version="1.3.1" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="3.4.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Scripting.Common" Version="3.4.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("CS-Script.Core", "1.3.1", "net8.0", [(null, [("Microsoft.CodeAnalysis.Scripting.Common", "[3.4.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Scripting.Common", "3.4.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[3.4.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "3.4.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("CS-Script.Core", "2.0.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Scripting.Common", "[3.6.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Scripting.Common", "3.6.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[3.6.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "3.6.0", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            // Add comment about root and dependencies
-            new Dependency("CS-Script.Core", "1.3.1", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.Common", "3.4.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.Scripting.Common", "3.4.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("CS-Script.Core", "2.0.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "CS-Script.Core/2.0.0",
-            "Microsoft.CodeAnalysis.Common/3.6.0",
-            "Microsoft.CodeAnalysis.Scripting.Common/3.6.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Updating a dependency (Microsoft.Bcl.AsyncInterfaces) of the root package (Azure.Core) will require the root package to also update, but since the dependency is not in the existing list, we do not include it
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewUpdatingNonExistingDependency(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Azure.Core" Version="1.21.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Azure.Core", "1.21.0", "net8.0", [(null, [("Microsoft.Bcl.AsyncInterfaces", "[1.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Bcl.AsyncInterfaces", "1.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Azure.Core", "1.22.0", "net8.0", [(null, [("Microsoft.Bcl.AsyncInterfaces", "[1.1.1]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Bcl.AsyncInterfaces", "1.1.1", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Azure.Core", "1.21.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.Bcl.AsyncInterfaces", "1.1.1", DependencyType.Unknown)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Azure.Core/1.22.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Adding a reference
-    // Newtonsoft.Json needs to update to 13.0.1. Although Newtonsoft.Json.Bson can use the original version of 12.0.1, for security vulnerabilities and
-    // because there is no later version of Newtonsoft.Json.Bson 1.0.2, Newtonsoft.Json would be added to the existing list to prevent resolution
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewUpdatingNonExistentDependencyAndKeepingReference(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Newtonsoft.Json.Bson" Version="1.0.2" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Newtonsoft.Json.Bson", "1.0.2", "net8.0", [(null, [("Newtonsoft.Json", "12.0.1")])]),
-            MockNuGetPackage.CreateSimplePackage("Newtonsoft.Json", "12.0.1", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Newtonsoft.Json", "13.0.1", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Newtonsoft.Json.Bson", "1.0.2", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Newtonsoft.Json", "13.0.1", DependencyType.Unknown)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Newtonsoft.Json.Bson/1.0.2",
-            "Newtonsoft.Json/13.0.1",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Updating unreferenced dependency
-    // Root package (Microsoft.CodeAnalysis.Compilers) and its dependencies (Microsoft.CodeAnalysis.CSharp), (Microsoft.CodeAnalysis.VisualBasic) are all 4.9.2
-    // These packages all require the transitive dependency of the root package (Microsoft.CodeAnalysis.Common) to be 4.9.2, but it's not in the existing list
-    // If Microsoft.CodeAnalysis.Common is updated to 4.10.0, everything else updates and Microsoft.CoseAnalysis.Common is not kept in the existing list
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewTransitiveDependencyNotExisting(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Microsoft.CodeAnalysis.Compilers" Version="4.9.2" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.9.2" />
-                <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic" Version="4.9.2" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Compilers", "4.9.2", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.9.2]"), ("Microsoft.CodeAnalysis.CSharp", "[4.9.2]"), ("Microsoft.CodeAnalysis.VisualBasic", "[4.9.2]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.9.2", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.9.2", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.VisualBasic", "4.9.2", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Compilers", "4.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.10.0]"), ("Microsoft.CodeAnalysis.CSharp", "[4.10.0]"), ("Microsoft.CodeAnalysis.VisualBasic", "[4.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.10.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.10.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.VisualBasic", "4.10.0", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Compilers", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.VisualBasic", "4.9.2", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.10.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Microsoft.CodeAnalysis.Compilers/4.10.0",
-            "Microsoft.CodeAnalysis.CSharp/4.10.0",
-            "Microsoft.CodeAnalysis.VisualBasic/4.10.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Updating referenced dependency
-    // The same as previous test, but the transitive dependency (Microsoft.CodeAnalysis.Common) is in the existing list
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewSingleTransitiveDependencyExisting(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Microsoft.CodeAnalysis.Compilers" Version="4.9.2" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.9.2" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.9.2" />
-                <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic" Version="4.9.2" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Compilers", "4.9.2", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.9.2]"), ("Microsoft.CodeAnalysis.CSharp", "[4.9.2]"), ("Microsoft.CodeAnalysis.VisualBasic", "[4.9.2]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.9.2", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.9.2", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.VisualBasic", "4.9.2", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Compilers", "4.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.10.0]"), ("Microsoft.CodeAnalysis.CSharp", "[4.10.0]"), ("Microsoft.CodeAnalysis.VisualBasic", "[4.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.10.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.10.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.VisualBasic", "4.10.0", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Compilers", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.VisualBasic", "4.9.2", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.10.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Microsoft.CodeAnalysis.Compilers/4.10.0",
-            "Microsoft.CodeAnalysis.Common/4.10.0",
-            "Microsoft.CodeAnalysis.CSharp/4.10.0",
-            "Microsoft.CodeAnalysis.VisualBasic/4.10.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // A combination of the third and fourth test, to measure efficiency of updating separate families
-    // Keeping a dependency that was not included in the original list (Newtonsoft.Json)
-    // Not keeping a dependency that was not included in the original list (Microsoft.CodeAnalysis.Common)
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewSelectiveAdditionPackages(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Microsoft.CodeAnalysis.Compilers" Version="4.9.2" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.9.2" />
-                <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic" Version="4.9.2" />
-                <PackageReference Include="Newtonsoft.Json.Bson" Version="1.0.2" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Compilers", "4.9.2", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.9.2]"), ("Microsoft.CodeAnalysis.CSharp", "[4.9.2]"), ("Microsoft.CodeAnalysis.VisualBasic", "[4.9.2]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.9.2", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.9.2", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.VisualBasic", "4.9.2", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Newtonsoft.Json.Bson", "1.0.2", "net8.0", [(null, [("Newtonsoft.Json", "13.0.1")])]),
-            MockNuGetPackage.CreateSimplePackage("Newtonsoft.Json", "13.0.1", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Compilers", "4.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.10.0]"), ("Microsoft.CodeAnalysis.CSharp", "[4.10.0]"), ("Microsoft.CodeAnalysis.VisualBasic", "[4.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.10.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.10.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.VisualBasic", "4.10.0", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Compilers", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.VisualBasic", "4.9.2", DependencyType.PackageReference),
-            new Dependency("Newtonsoft.Json.Bson", "1.0.2", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.10.0", DependencyType.PackageReference),
-            new Dependency("Newtonsoft.Json", "13.0.1", DependencyType.Unknown)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Microsoft.CodeAnalysis.Compilers/4.10.0",
-            "Microsoft.CodeAnalysis.CSharp/4.10.0",
-            "Microsoft.CodeAnalysis.VisualBasic/4.10.0",
-            "Newtonsoft.Json.Bson/1.0.2",
-            "Newtonsoft.Json/13.0.1",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Two top level packages (Buildalyzer), (Microsoft.CodeAnalysis.CSharp.Scripting) that share a dependency (Microsoft.CodeAnalysis.Csharp)
-    // Updating ONE of the top level packages, which updates the dependencies and their other "parents"
-    // First family: Buildalyzer 7.0.1 requires Microsoft.CodeAnalysis.CSharp to be = 4.0.1 and Microsoft.CodeAnalysis.Common to be 4.0.1 (@ 6.0.4, Microsoft.CodeAnalysis.Common isn't a dependency of buildalyzer)
-    // Second family: Microsoft.CodeAnalysis.CSharp.Scripting 4.0.1 requires Microsoft.CodeAnalysis.CSharp 4.0.1 and Microsoft.CodeAnalysis.Common to be 4.0.1 (Specific version)
-    // Updating Buildalyzer to 7.0.1 will update its transitive dependency (Microsoft.CodeAnalysis.Common) and then its transitive dependency's "family"
-    [Fact]
-    public async Task DependencyConflictsCanBeResolvedNewSharingDependency()
-    {
-        // arrange
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Buildalyzer" Version="6.0.4" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Csharp.Scripting" Version="3.10.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.10.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="3.10.0" />
-              </ItemGroup>
-            </Project>
-            """);
-
-        var testPackages = new MockNuGetPackage[]
-        {
-            MockNuGetPackage.CreateSimplePackage("Buildalyzer", "6.0.4", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[3.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Buildalyzer", "7.0.1", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.0.1]")])]),
-
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Scripting", "3.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[3.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Scripting", "4.0.1", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.0.1]")])]),
-
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "3.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[3.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.0.1", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.0.1]")])]),
-
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "3.10.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.0.1", "net8.0"),
-        };
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory(testPackages, tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Buildalyzer", "6.0.4", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp.Scripting", "3.10.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp", "3.10.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.Common", "3.10.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Buildalyzer", "7.0.1", DependencyType.PackageReference),
-        }.ToImmutableArray();
-
-        // act
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager(),
-            new TestLogger()
-        );
-
-        // assert
-        var expectedDependencies = new[]
-        {
-            "Buildalyzer/7.0.1",
-            "Microsoft.CodeAnalysis.CSharp.Scripting/4.0.1",
-            "Microsoft.CodeAnalysis.CSharp/4.0.1",
-            "Microsoft.CodeAnalysis.Common/4.0.1"
-        };
-        Assert.NotNull(resolvedDependencies);
-        var actualDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        AssertEx.Equal(expectedDependencies, actualDependencies);
-    }
-
-    // Updating two families at once to test efficiency
-    // First family: Direct dependency (Microsoft.CodeAnalysis.Common) needs to be updated, which will then need to update in the existing list its dependency (System.Collections.Immutable) and "parent" (Microsoft.CodeAnalysis.Csharp.Scripting)
-    // Second family: Updating the root package (Azure.Core) in the existing list will also need to update its dependency (Microsoft.Bcl.AsyncInterfaces)
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewUpdatingEntireFamily(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="System.Collections.Immutable" Version="7.0.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.8.0" />
-                <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="1.0.0" />
-                <PackageReference Include="Azure.Core" Version="1.21.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Scripting", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.8.0]"), ("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.8.0", "net8.0", [(null, [("System.Collections.Immutable", "7.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "7.0.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Azure.Core", "1.21.0", "net8.0", [(null, [("Microsoft.Bcl.AsyncInterfaces", "1.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Bcl.AsyncInterfaces", "1.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Scripting", "4.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.10.0]"), ("Microsoft.CodeAnalysis.Common", "[4.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.10.0", "net8.0", [(null, [("System.Collections.Immutable", "8.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "8.0.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Azure.Core", "1.22.0", "net8.0", [(null, [("Microsoft.Bcl.AsyncInterfaces", "1.1.1")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Bcl.AsyncInterfaces", "1.1.1", "net8.0"),
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("System.Collections.Immutable", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp.Scripting", "4.8.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.Bcl.AsyncInterfaces", "1.0.0", DependencyType.PackageReference),
-            new Dependency("Azure.Core", "1.21.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.10.0", DependencyType.PackageReference),
-            new Dependency("Azure.Core", "1.22.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "System.Collections.Immutable/8.0.0",
-            "Microsoft.CodeAnalysis.CSharp.Scripting/4.10.0",
-            "Microsoft.Bcl.AsyncInterfaces/1.1.1",
-            "Azure.Core/1.22.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Similar to the last test, except Microsoft.CodeAnalysis.Common is in the existing list
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewUpdatingTopLevelAndDependency(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="System.Collections.Immutable" Version="7.0.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.8.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.8.0" />
-                <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="1.0.0" />
-                <PackageReference Include="Azure.Core" Version="1.21.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Scripting", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.8.0]"), ("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.8.0", "net8.0", [(null, [("System.Collections.Immutable", "7.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "7.0.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Azure.Core", "1.21.0", "net8.0", [(null, [("Microsoft.Bcl.AsyncInterfaces", "1.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Bcl.AsyncInterfaces", "1.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Scripting", "4.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.10.0]"), ("Microsoft.CodeAnalysis.Common", "[4.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.10.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.10.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.10.0", "net8.0", [(null, [("System.Collections.Immutable", "8.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "8.0.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Azure.Core", "1.22.0", "net8.0", [(null, [("Microsoft.Bcl.AsyncInterfaces", "1.1.1")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Bcl.AsyncInterfaces", "1.1.1", "net8.0"),
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("System.Collections.Immutable", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp.Scripting", "4.8.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.8.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.Bcl.AsyncInterfaces", "1.0.0", DependencyType.Unknown),
-            new Dependency("Azure.Core", "1.21.0", DependencyType.PackageReference),
-
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.10.0", DependencyType.PackageReference),
-            new Dependency("Azure.Core", "1.22.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "System.Collections.Immutable/8.0.0",
-            "Microsoft.CodeAnalysis.CSharp.Scripting/4.10.0",
-            "Microsoft.CodeAnalysis.Common/4.10.0",
-            "Microsoft.Bcl.AsyncInterfaces/1.1.1",
-            "Azure.Core/1.22.0"
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Out of scope test: AutoMapper.Extensions.Microsoft.DependencyInjection's versions are not yet compatible
-    // To update root package (AutoMapper.Collection) to 10.0.0, its dependency (AutoMapper) needs to update to 13.0.0. 
-    // However, there is no higher version of AutoMapper's other "parent" (AutoMapper.Extensions.Microsoft.DependencyInjection) that is compatible with the new version
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewOutOfScope(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
-                <PackageReference Include="AutoMapper" Version="12.0.1" />
-                <PackageReference Include="AutoMapper.Collection" Version="9.0.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            MockNuGetPackage.CreateSimplePackage("AutoMapper.Extensions.Microsoft.DependencyInjection", "12.0.1", "net8.0", [(null, [("AutoMapper", "[12.0.1]")])]),
-            MockNuGetPackage.CreateSimplePackage("AutoMapper", "12.0.1", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("AutoMapper", "13.0.1", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("AutoMapper.Collection", "9.0.0", "net8.0", [(null, [("AutoMapper", "[12.0.0, 13.0.0)")])]),
-            MockNuGetPackage.CreateSimplePackage("AutoMapper.Collection", "10.0.0", "net8.0", [(null, [("AutoMapper", "[13.0.0, 14.0.0)")])])
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("AutoMapper.Extensions.Microsoft.DependencyInjection", "12.0.1", DependencyType.PackageReference),
-            new Dependency("AutoMapper", "12.0.1", DependencyType.PackageReference),
-            new Dependency("AutoMapper.Collection", "9.0.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("AutoMapper.Collection", "10.0.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "AutoMapper.Extensions.Microsoft.DependencyInjection/12.0.1",
-            "AutoMapper/12.0.1",
-            "AutoMapper.Collection/9.0.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Two dependencies (Microsoft.Extensions.Caching.Memory), (Microsoft.EntityFrameworkCore.Analyzers) used by the same parent (Microsoft.EntityFrameworkCore), updating one of the dependencies
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewTwoDependenciesShareSameParent(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.11" />
-                <PackageReference Include="Microsoft.EntityFrameworkCore.Analyzers" Version="7.0.11" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore", "7.0.11", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Analyzers", "7.0.11"), ("Microsoft.Extensions.Caching.Memory", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Analyzers", "7.0.11", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Extensions.Caching.Memory", "7.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore", "8.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Analyzers", "8.0.0"), ("Microsoft.Extensions.Caching.Memory", "[8.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Analyzers", "8.0.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.Extensions.Caching.Memory", "8.0.0", "net8.0"),
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Microsoft.EntityFrameworkCore", "7.0.11", DependencyType.PackageReference),
-            new Dependency("Microsoft.EntityFrameworkCore.Analyzers", "7.0.11", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.Extensions.Caching.Memory", "8.0.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Microsoft.EntityFrameworkCore/8.0.0",
-            "Microsoft.EntityFrameworkCore.Analyzers/8.0.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Updating referenced package
-    // 4 dependency chain to be updated. Since the package to be updated is in the existing list, do not update its parents since we want to change as little as possible
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewFamilyOfFourExisting(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.0" />
-                <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.0" />
-                <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.0" />
-                <PackageReference Include="Microsoft.EntityFrameworkCore.Analyzers" Version="7.0.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Design", "7.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Relational", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Relational", "7.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore", "7.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Analyzers", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Analyzers", "7.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Analyzers", "8.0.0", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Microsoft.EntityFrameworkCore.Design", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.EntityFrameworkCore.Relational", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.EntityFrameworkCore", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.EntityFrameworkCore.Analyzers", "7.0.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.EntityFrameworkCore.Analyzers", "8.0.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Microsoft.EntityFrameworkCore.Design/7.0.0",
-            "Microsoft.EntityFrameworkCore.Relational/7.0.0",
-            "Microsoft.EntityFrameworkCore/7.0.0",
-            "Microsoft.EntityFrameworkCore.Analyzers/8.0.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Updating unreferenced package
-    // 4 dependency chain to be updated, dependency to be updated is not in the existing list, so its family will all be updated
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewFamilyOfFourNotInExisting(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.0" />
-                <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.0" />
-                <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Design", "7.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Relational", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Relational", "7.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore", "7.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Analyzers", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Analyzers", "7.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Design", "8.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Relational", "[8.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Relational", "8.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore", "[8.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore", "8.0.0", "net8.0", [(null, [("Microsoft.EntityFrameworkCore.Analyzers", "[8.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.EntityFrameworkCore.Analyzers", "8.0.0", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Microsoft.EntityFrameworkCore.Design", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.EntityFrameworkCore.Relational", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.EntityFrameworkCore", "7.0.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Microsoft.EntityFrameworkCore.Analyzers", "8.0.0", DependencyType.PackageReference)
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Microsoft.EntityFrameworkCore.Design/8.0.0",
-            "Microsoft.EntityFrameworkCore.Relational/8.0.0",
-            "Microsoft.EntityFrameworkCore/8.0.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Updating a referenced transitive dependency
-    // Updating a transtitive dependency (System.Collections.Immutable) to 8.0.0, which will update its "parent" (Microsoft.CodeAnalysis.CSharp) and its "grandparent" (Microsoft.CodeAnalysis.CSharp.Workspaces) to update
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewFamilyOfFourSpecificExisting(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="System.Collections.Immutable" Version="7.0.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.8.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.8.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "7.0.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Workspaces", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.8.0]"), ("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.8.0", "net8.0", [(null, [("System.Collections.Immutable", "7.0.0")])]),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "8.0.0", "net8.0")
-        ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("System.Collections.Immutable", "7.0.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp.Workspaces", "4.8.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp", "4.8.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.8.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("System.Collections.Immutable", "8.0.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "System.Collections.Immutable/8.0.0",
-            "Microsoft.CodeAnalysis.CSharp.Workspaces/4.8.0",
-            "Microsoft.CodeAnalysis.CSharp/4.8.0",
-            "Microsoft.CodeAnalysis.Common/4.8.0",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    // Similar to the last test, with the "grandchild" (System.Collections.Immutable) not in the existing list
-    [Theory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task DependencyConflictsCanBeResolvedNewFamilyOfFourSpecificNotInExisting(bool useExistingSdks)
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.8.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
-                <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.8.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Workspaces", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.8.0]"), ("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.8.0", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.8.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.8.0", "net8.0", [(null, [("System.Collections.Immutable", "[7.0.0]")])]),
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "7.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("System.Collections.Immutable", "8.0.0", "net8.0"),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp.Workspaces", "4.9.2", "net8.0", [(null, [("Microsoft.CodeAnalysis.CSharp", "[4.9.2]"), ("Microsoft.CodeAnalysis.Common", "[4.9.2]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.CSharp", "4.9.2", "net8.0", [(null, [("Microsoft.CodeAnalysis.Common", "[4.9.2]")])]),
-            MockNuGetPackage.CreateSimplePackage("Microsoft.CodeAnalysis.Common", "4.9.2", "net8.0", [(null, [("System.Collections.Immutable", "[8.0.0]")])]),
-       ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Microsoft.CodeAnalysis.CSharp.Workspaces", "4.8.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.CSharp", "4.8.0", DependencyType.PackageReference),
-            new Dependency("Microsoft.CodeAnalysis.Common", "4.8.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("System.Collections.Immutable", "8.0.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = useExistingSdks },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Microsoft.CodeAnalysis.CSharp.Workspaces/4.9.2",
-            "Microsoft.CodeAnalysis.CSharp/4.9.2",
-            "Microsoft.CodeAnalysis.Common/4.9.2",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    [Fact(Timeout = 120_000)] // 2m
-    public async Task DependencyConflictsCanBeResolved_TopLevelDependencyHasNewerVersionsThatDoNotPullUpTransitive()
-    {
-        using var tempDirectory = new TemporaryDirectory();
-        var projectPath = Path.Join(tempDirectory.DirectoryPath, "project.csproj");
-        await File.WriteAllTextAsync(projectPath, """
-            <Project Sdk="Microsoft.NET.Sdk">
-              <PropertyGroup>
-                <TargetFramework>net8.0</TargetFramework>
-              </PropertyGroup>
-              <ItemGroup>
-                <PackageReference Include="Top.Level.Package" Version="1.41.0" />
-              </ItemGroup>
-            </Project>
-            """);
-        await UpdateWorkerTestBase.MockNuGetPackagesInDirectory([
-            // initial packages
-            MockNuGetPackage.CreateSimplePackage("Top.Level.Package", "1.41.0", "net8.0", [(null, [("Transitive.Package", "6.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("Transitive.Package", "6.0.0", "net8.0"),
-            // available packages
-            MockNuGetPackage.CreateSimplePackage("Top.Level.Package", "1.45.0", "net8.0", [(null, [("Transitive.Package", "6.0.0")])]),
-            MockNuGetPackage.CreateSimplePackage("Transitive.Package", "8.0.5", "net8.0"),
-       ], tempDirectory.DirectoryPath);
-
-        var dependencies = new[]
-        {
-            new Dependency("Top.Level.Package", "1.41.0", DependencyType.PackageReference),
-        }.ToImmutableArray();
-        var update = new[]
-        {
-            new Dependency("Transitive.Package", "8.0.5", DependencyType.PackageReference),
-        }.ToImmutableArray();
-
-        var resolvedDependencies = await MSBuildHelper.ResolveDependencyConflicts(
-            tempDirectory.DirectoryPath,
-            projectPath,
-            "net8.0",
-            dependencies,
-            update,
-            new ExperimentsManager() { InstallDotnetSdks = true },
-            new TestLogger()
-        );
-        Assert.NotNull(resolvedDependencies);
-        var actualResolvedDependencies = resolvedDependencies.Value.Select(d => $"{d.Name}/{d.Version}").ToArray();
-        var expectedResolvedDependencies = new[]
-        {
-            "Top.Level.Package/1.41.0",
-            "Transitive.Package/8.0.5",
-        };
-        AssertEx.Equal(expectedResolvedDependencies, actualResolvedDependencies);
-    }
-
-    #endregion
-
     [Theory]
     [MemberData(nameof(GenerateErrorFromToolOutputTestData))]
     public async Task GenerateErrorFromToolOutput(string output, JobErrorBase? expectedError)
